Enhance onetime config handling by adding input hash management. Updated OnetimeConfigInfoManager to track input hashes and modified related classes to support new logic for determining config status and expiration. Adjusted CollectionPipeline and CollectionPipelineManager to utilize input hashes during config updates. Improved unit tests to cover new functionality.

+++ core/file_server/StaticFileServer.cpp
@@ -78,7 +78,7 @@ void StaticFileServer::ClearUnusedCheckpoints() {
     mIsUnusedCheckpointsCleared = true;
 }
 
-void StaticFileServer::RemoveInput(const string& configName, size_t idx) {
+void StaticFileServer::RemoveInput(const string& configName, size_t idx, bool shouldDeleteCheckpoint) {
     {
         lock_guard<mutex> lock(mUpdateMux);
         mInputFileDiscoveryConfigsMap.erase(make_pair(configName, idx));
@@ -87,7 +87,9 @@ void StaticFileServer::RemoveInput(const string& configName, size_t idx) {
         mInputFileTagConfigsMap.erase(make_pair(configName, idx));
         mDeletedInputs.emplace(configName, idx);
     }
-    InputStaticFileCheckpointManager::GetInstance()->DeleteCheckpoint(configName, idx);
+    if (shouldDeleteCheckpoint) {
+        InputStaticFileCheckpointManager::GetInstance()->DeleteCheckpoint(configName, idx);
+    }
 }
 
 void StaticFileServer::AddInput(const string& configName,
